requirements.txt
aiogram~=2.20
Pillow~=9.1.1
pytz~=2022.1